requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
numpy<2
sentence-transformers
faiss-cpu
pymupdf
groq
mcp
langchain
langchain-community
duckduckgo-search
ddgs
beautifulsoup4
